+++ bioblend/galaxy/__init__.py
@@ -38,6 +38,7 @@ class GalaxyInstance(GalaxyClient):
         key: Optional[str] = None,
         email: Optional[str] = None,
         password: Optional[str] = None,
+        *,
         verify: bool = True,
     ) -> None:
         """
@@ -81,7 +82,7 @@ class GalaxyInstance(GalaxyClient):
         :param verify: Whether to verify the server's TLS certificate
         :type verify: bool
         """
-        super().__init__(url, key, email, password, verify=verify)
+        super().__init__(url, key=key, email=email, password=password, verify=verify)
         self.libraries = libraries.LibraryClient(self)
         self.histories = histories.HistoryClient(self)
         self.workflows = workflows.WorkflowClient(self)